add the database files copied from prjxray run to this repo
These should be updateable from the master run without any harm,
but I also don't anticipate them ever changing.

Note that these were generated by running the prjxray fuzzer specifically
on the part number used by Betrusted, so the database is basically
the minimum necessary information needed to map the INIT values to
LUTs in the ROM as-designed.

If more features are necessary, then the database files will need
to be updated. However, as long as prjxray does not do a major
change to the database format, this should be a fairly transparent
process.