change byte ordering of ROM binary to LUT layout from little to big
crytpographic material tends to be big-endian, so this leads to
a more "expected" result.

Also add the beginnings of the Rust code generator and some comment
fixes; realized that LUT ordering is /not/ dependent on the rom.db
ordering, because the BELs are resolved by name to bit position
inside the code. So even if the LUT order is permuted in rom.db,
the BEL-to-bank ordering is preserved.