refactor key2bits to automatically translate frame addresses
The original program output frame addresses for patching, and
required doing a "diff" of two bitstreams with different
data in the ROM to figure out the relative position of the patch.

Now, we read in the part file, which contains the mapping of
frames to bitstream positions. The function address_to_framestream()
can use this database to translate an arbitrary frame address
to a relative address in the bitstream, suitable for patching.